feat: refactor rating modal - move form to a separate component
