+++ src/Components/FeedbackForm/types.ts
@@ -1,0 +1,3 @@
+export type FeedbackFormPropTypes = {
+  onClose: () => void;
+};
