+++ src/Components/RatingModal/styles.scss
@@ -6,54 +6,4 @@
   &__heading {
     text-align: center;
   }
-
-  &__form {
-    box-sizing: border-box;
-    overflow: hidden;
-    margin-top: 1em;
-    display: flex;
-    flex-direction: column;
-    gap: 1em;
-
-    &--section {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      label {
-        align-self: flex-start;
-      }
-    }
-
-    &--textarea {
-      width: 90%;
-      margin-top: 1em;
-      box-sizing: border-box;
-      padding: 0.5em;
-      font-size: 16px;
-      border-radius: 8px;
-      outline: none;
-      border: 2px solid transparent;
-      resize: none;
-      &:focus {
-        border: 2px solid #ea850f;
-      }
-    }
-
-    &--submit {
-      cursor: pointer;
-      width: 100%;
-      padding: 10px 30px;
-      border: 3px solid #131722;
-      border-radius: 8px;
-      background: linear-gradient(to right, #ea850f, #ffc700);
-      color: #000;
-      font-size: 14px;
-      font-weight: 700;
-      outline: none;
-      &:disabled {
-        background: #808080;
-        cursor: default;
-      }
-    }
-  }
 }
